worker/requirements.txt
# ==============================================================================
# Worker Service Dependencies
# ==============================================================================
# This file contains Python dependencies for the Airflow worker service.
# Note: FastAPI/uvicorn are NOT included here - API is now served by Next.js
# ==============================================================================

# Airflow DAG dependencies
requests
beautifulsoup4
mysqlclient
sqlalchemy
pydantic
pydantic-settings
python-dotenv
pymysql
pendulum
selenium

# paperprocessor dependencies
pillow
numpy
PyMuPDF
scipy
rapidfuzz
openai
mistralai
aiofiles
pdf2image

# shared module dependencies
qdrant-client
httpx
aiohttp
email-validator

# Fix protobuf version conflicts
protobuf==3.20.3
